Refactor: Make `RemoteFeatureFlagStoreMock` accessible from other tests
