+++ WordPress/WordPressTest/JetpackBrandingMenuCardPresenterTests.swift
@@ -114,35 +114,6 @@ final class JetpackBrandingMenuCardPresenterTests: XCTestCase {
 
 }
 
-private class RemoteFeatureFlagStoreMock: RemoteFeatureFlagStore {
-
-    var removalPhaseOne = false
-    var removalPhaseTwo = false
-    var removalPhaseThree = false
-    var removalPhaseFour = false
-    var removalPhaseNewUsers = false
-
-    override func value(for flag: OverrideableFlag) -> Bool {
-        guard let flag = flag as? WordPress.FeatureFlag else {
-            return false
-        }
-        switch flag {
-        case .jetpackFeaturesRemovalPhaseOne:
-            return removalPhaseOne
-        case .jetpackFeaturesRemovalPhaseTwo:
-            return removalPhaseTwo
-        case .jetpackFeaturesRemovalPhaseThree:
-            return removalPhaseThree
-        case .jetpackFeaturesRemovalPhaseFour:
-            return removalPhaseFour
-        case .jetpackFeaturesRemovalPhaseNewUsers:
-            return removalPhaseNewUsers
-        default:
-            return super.value(for: flag)
-        }
-    }
-}
-
 private class RemoteConfigStoreMock: RemoteConfigStore {
 
     var phaseThreeBlogPostUrl: String?
