+++ WordPress/WordPressTest/JetpackNotificationMigrationServiceTests.swift
@@ -137,15 +137,17 @@ private class RemoteNotificationRegisterMock: RemoteNotificationRegister {
     }
 }
 
-private class RemoteFeatureFlagStoreMock: RemoteFeatureFlagStore {
-    var value = false {
-        didSet {
-            try? FeatureFlagOverrideStore().override(FeatureFlag.jetpackMigrationPreventDuplicateNotifications, withValue: value)
+private extension JetpackNotificationMigrationServiceTests {
+    class RemoteFeatureFlagStoreMock: RemoteFeatureFlagStore {
+        var value = false {
+            didSet {
+                try? FeatureFlagOverrideStore().override(FeatureFlag.jetpackMigrationPreventDuplicateNotifications, withValue: value)
+            }
         }
-    }
 
-    override func value(for flag: OverrideableFlag) -> Bool {
-        return value
+        override func value(for flag: OverrideableFlag) -> Bool {
+            return value
+        }
     }
 }
